apps/graphiti-service/pyproject.toml
[project]
name = "graphiti-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "graphiti-core[kuzu]",
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"
